renamed TrackedActivation to Nonlinear
