+++ alpine/models/nonlin/nonlin.py
@@ -3,9 +3,20 @@ import numpy as np
 
 import torch.nn as nn
 
-from alpine.models.nonlin.tracked_activation import TrackedActivation
+import torch
+import torch.nn as nn
 
-class Sine(TrackedActivation):
+class Nonlinear(nn.Module):
+    """Base class for activation functions that can be tracked using the FeatureExtractor context manager."""
+    def __init__(self, name: str):
+        super().__init__()
+        self.name = name.lower()
+
+    def forward(self, x):
+        raise NotImplementedError("Subclasses must implement forward()")
+    
+    
+class Sine(Nonlinear):
     """Sine nonlinearity proposed by :cite:`siren2020sitzmann`"""
     def __init__(self, omega=30.0, name='sine'):
         super().__init__(name=name)
@@ -15,7 +26,7 @@ class Sine(TrackedActivation):
     def forward(self, x):
         return torch.sin(self.omega * x)
 
-class ReLU(TrackedActivation):
+class ReLU(Nonlinear):
     def __init__(self, name='relu'):
         super().__init__(name=name)
         self.relu = nn.ReLU()
@@ -24,7 +35,7 @@ class ReLU(TrackedActivation):
     def forward(self, x):
         return self.relu(x)
 
-class Gauss(TrackedActivation):
+class Gauss(Nonlinear):
     '''Gaussian nonlinearity proposed by :cite:`ramasinghe2022beyond`. '''
     def __init__(self, scale=1.0, name='gauss'):
         super().__init__(name=name)
@@ -35,7 +46,7 @@ class Gauss(TrackedActivation):
         return torch.exp(-(self.scale * x)**2)
 
 
-class Wavelet(TrackedActivation):
+class Wavelet(Nonlinear):
     ''' Wavelet nonlinearty proposed by :cite:`saragadam2022wire` '''
     def __init__(self, sigma=1.0, omega=30.0, trainable=False, name="wavelet"):
         super(Wavelet, self).__init__(name=name)
@@ -47,7 +58,7 @@ class Wavelet(TrackedActivation):
         return torch.exp(1j * (self.omega * x)  - (self.sigma * x).abs().square())
 
 
-class HOSC(TrackedActivation):
+class HOSC(Nonlinear):
     '''HOSC nonlinearity proposed by :cite:`serrano2024hoscperiodicactivationfunction` '''
     def __init__(self, beta, name='hosc'):
         super().__init__(name=name)
@@ -57,7 +68,7 @@ class HOSC(TrackedActivation):
     def forward(self, x):
         return torch.tanh(self.beta * torch.sin(x))
 
-class Sinc(TrackedActivation):
+class Sinc(Nonlinear):
     '''Sinc nonlinearity proposed by :cite:``. '''
     def __init__(self, omega=30.0, name='sinc'):
         super().__init__(name=name)
